refactor: Move get_embedding_model to models folder
