+++ jobs/ml_jobs/edito_semantic_search/app/search/client.py
@@ -5,7 +5,8 @@ import pyarrow.dataset as ds
 from loguru import logger
 from pyarrow.fs import GcsFileSystem
 
-from app.constants import PARQUET_FILE, get_embedding_model
+from app.constants import PARQUET_FILE
+from app.models.embedding import get_embedding_model
 from app.search.filters import apply_filters
 
 PARTITION_COLS = ["offer_subcategory_id"]
